huge refactor, maybe revert?
