+++ src/ytdl_sub/config/overrides.py
@@ -5,7 +5,6 @@ from typing import Optional
 from typing import Set
 
 import mergedeep
-from yt_dlp.utils import sanitize_filename
 
 from ytdl_sub.entries.entry import Entry
 from ytdl_sub.entries.script.variable_definitions import VARIABLES
