+++ src/ytdl_sub/config/plugin/plugin.py
@@ -8,7 +8,7 @@ from typing import Tuple
 from typing import Type
 
 from ytdl_sub.config.overrides import Overrides
-from ytdl_sub.config.preset_options import TOptionsValidator
+from ytdl_sub.config.validators.options import TOptionsValidator
 from ytdl_sub.entries.entry import Entry
 from ytdl_sub.utils.file_handler import FileMetadata
 from ytdl_sub.ytdl_additions.enhanced_download_archive import DownloadArchiver
