+++ src/ytdl_sub/config/preset_options.py
@@ -1,13 +1,7 @@
-from abc import ABC
 from typing import Any
-from typing import Dict
 from typing import Optional
-from typing import Set
-from typing import TypeVar
 
 from ytdl_sub.config.defaults import DEFAULT_DOWNLOAD_ARCHIVE_NAME
-from ytdl_sub.config.plugin_operation import PluginOperation
-from ytdl_sub.utils.exceptions import ValidationException
 from ytdl_sub.validators.file_path_validators import OverridesStringFormatterFilePathValidator
 from ytdl_sub.validators.file_path_validators import StringFormatterFileNameValidator
 from ytdl_sub.validators.strict_dict_validator import StrictDictValidator
@@ -17,55 +11,6 @@ from ytdl_sub.validators.string_formatter_validators import OverridesStringForma
 from ytdl_sub.validators.string_formatter_validators import StringFormatterValidator
 from ytdl_sub.validators.validators import BoolValidator
 from ytdl_sub.validators.validators import LiteralDictValidator
-from ytdl_sub.validators.validators import Validator
-
-
-# pylint: disable=no-self-use
-# pylint: disable=unused-argument
-class OptionsValidator(Validator, ABC):
-    """
-    Abstract class that validates options for preset sections (plugins, downloaders)
-    """
-
-    def validation_exception(
-        self,
-        error_message: str | Exception,
-    ) -> ValidationException:
-        """
-        Parameters
-        ----------
-        error_message
-            Error message to include in the validation exception
-
-        Returns
-        -------
-        Validation exception that points to the location in the config. To be used to throw good
-        validation exceptions at runtime from code outside this class.
-        """
-        return self._validation_exception(error_message=error_message)
-
-    def added_source_variables(
-        self, unresolved_variables: Set[str]
-    ) -> Dict[PluginOperation, Set[str]]:
-        """
-        If the plugin adds source variables, list them here.
-
-        Returns
-        -------
-        List of added source variables this plugin creates
-        """
-        return {}
-
-
-TOptionsValidator = TypeVar("TOptionsValidator", bound=OptionsValidator)
-
-
-class OptionsDictValidator(StrictDictValidator, OptionsValidator, ABC):
-    pass
-
-
-# pylint: enable=no-self-use
-# pylint: enable=unused-argument
 
 
 class YTDLOptions(LiteralDictValidator):
