+++ src/ytdl_sub/plugins/audio_extract.py
@@ -4,9 +4,9 @@ from typing import Dict
 from typing import Optional
 from typing import Set
 
-from ytdl_sub.config.plugin import Plugin
-from ytdl_sub.config.plugin_operation import PluginOperation
-from ytdl_sub.config.preset_options import OptionsDictValidator
+from ytdl_sub.config.plugin.plugin import Plugin
+from ytdl_sub.config.plugin.plugin_operation import PluginOperation
+from ytdl_sub.config.validators.options import OptionsDictValidator
 from ytdl_sub.downloaders.ytdl_options_builder import YTDLOptionsBuilder
 from ytdl_sub.entries.entry import Entry
 from ytdl_sub.entries.script.variable_definitions import VARIABLES as v
@@ -68,9 +68,7 @@ class AudioExtractOptions(OptionsDictValidator):
             return self._quality.value
         return None
 
-    def added_source_variables(
-        self, unresolved_variables: Set[str]
-    ) -> Dict[PluginOperation, Set[str]]:
+    def modified_variables(self) -> Dict[PluginOperation, Set[str]]:
         """
         Possibly changes ``ext``, so do not resolve until this has run
         """
