+++ src/ytdl_sub/plugins/regex.py
@@ -6,9 +6,9 @@ from typing import Optional
 from typing import Set
 
 from ytdl_sub.config.overrides import Overrides
-from ytdl_sub.config.plugin import Plugin
-from ytdl_sub.config.plugin_operation import PluginOperation
-from ytdl_sub.config.preset_options import OptionsDictValidator
+from ytdl_sub.config.plugin.plugin import Plugin
+from ytdl_sub.config.plugin.plugin_operation import PluginOperation
+from ytdl_sub.config.validators.options import OptionsDictValidator
 from ytdl_sub.entries.entry import Entry
 from ytdl_sub.script.parser import parse
 from ytdl_sub.script.utils.exceptions import ScriptVariableNotResolved
