+++ src/ytdl_sub/subscriptions/base_subscription.py
@@ -5,7 +5,7 @@ from typing import Optional
 from ytdl_sub.config.config_validator import ConfigOptions
 from ytdl_sub.config.overrides import Overrides
 from ytdl_sub.config.preset import Preset
-from ytdl_sub.config.preset import PresetPlugins
+from ytdl_sub.config.plugin.preset_plugins import PresetPlugins
 from ytdl_sub.config.preset_options import OutputOptions
 from ytdl_sub.config.preset_options import YTDLOptions
 from ytdl_sub.downloaders.url.validators import MultiUrlValidator
